backend/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.4
bcrypt==3.1.7
beautifulsoup4==4.9.1
blinker==1.4
Brlapi==0.7.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
crcmod==1.7
cryptography==3.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.5.0
distro-info===0.23ubuntu1
Django==3.2.4
django-cors-headers==3.10.0
djangorestframework==3.12.4
duplicity==0.8.12.0
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gyp==0.1
html5lib==1.1
httplib2==0.18.1
idna==2.10
importlib-metadata==1.6.0
jeepney==0.4.3
keyring==21.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
louis==3.14.0
lxml==4.5.2
macaroonbakery==1.3.1
Mako==1.1.2
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.18.4
oauthlib==3.1.0
olefile==0.46
openshot-qt==2.5.1
paramiko==2.7.1
pexpect==4.6.0
Pillow==7.2.0
pipenv==11.9.0
protobuf==3.12.3
pyactiveresource==2.2.2
pycairo==1.16.2
pycups==2.0.1
PyGObject==3.38.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.4.0
PyQt5==5.15.0
pyRFC3339==1.1
python-apt==2.1.3+ubuntu1.4
python-dateutil==2.8.1
python-debian==0.1.37
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.2
reportlab==3.5.47
requests==2.23.0
requests-unixsocket==0.2.0
scour==0.37
SecretStorage==3.1.2
ShopifyAPI==9.0.0
simplejson==3.17.0
sip==4.19.24
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
systemd-python==234
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.7
virtualenv==20.0.29+ds
virtualenv-clone==0.3.0
wadllib==1.3.4
webencodings==0.5.1
xkit==0.0.0
zipp==1.0.0
